rename and fix

+++ lib/sprockets-digest-assets-fix.rb
@@ -7,7 +7,11 @@ module Sprockets
         full_digest_path = File.join dir, digest_path
         full_non_digest_path = File.join dir, info['logical_path']
         logger.info "Writing #{full_non_digest_path}"
-        FileUtils.cp full_digest_path, full_non_digest_path
+        if File.exists?(full_digest_path)
+          FileUtils.cp full_digest_path, full_non_digest_path
+        else
+          logger.info "Not found: #{full_digest_path}"
+        end
       end
     end
